add abstract table meta handler.
